pyproject.toml
[tool.poetry]
name = "vinny"
version = "0.1.0"
description = "An app for transcription."
authors = ["Greg", "Nick"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
six = "^1.16.0"
google-cloud-speech = "^2.20.0"
flask = "^2.3.2"
flask-socketio = "^5.3.4"
word2number = "^1.1"
openai = "^0.27.8"
python-socketio = "^5.8.0"
pyaudio = "^0.2.13"
eventlet = "^0.33.3"
requests = "^2.31.0"
websocket-client = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
mypy = "^1.4.1"
blue = "^0.9.1"
isort = "^5.12.0"
pylint = "^2.17.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.mypy]
exclude = ['venv', '.venv', 'scratch']